Changed explicity CLI/SEI to ATOMIC_BLOCK
The original intent of the cli/sei in dht_read was to block intterupts
during a critical timing section. This assumes interrupts are being used
ni the microcontrollers application. If interrupts were never used, this
libary enabled them, even though it does not use interrupts. A better
solution is to use avr-libc ATOMIC_BLOCK(ATOMIC_RESTORESTATE) which will
ensure interrupts are not enabled where they are not intended