Removed potential bus conflict
Bus was being held high by MCU when the DHT could possibly pull it low.
If this happens the DHT is fighting a Hi-Z state on the bus putting it
in an indeterminate state. Instead the bus is now released by setting
the bus pin to input, instead of driving the pin high/waiting/then
putting the pin to input.