Started making multiple sensors work on Node
Currently is not ready to be tested or pulled to master!!!

Began on making necessary changes for light, dht, and temperature
sensors to work at the same time

Trying to make it possible to have up to 2 light sensors, 4 dht sensors,
and 4 temperature sensors on one Sensor Node.

Note: Main contains lots of hard-coded indexes. This will be removed ASAP.
Index 0 will be the first light sensor and for the purposes of quick compilation
and testing some of these changes were temporarily necessary but may dramatically
change later.