Lots of small changes, needs more polish
No longer using an array of parser_flags, should have known that was a bad
idea.

Now uses the measure_all flag for READ

READ is currently the only one which prints things out exactly in the format it
should. ie: Lx=val Lx=val DTx=val Tx=val ...  where x is the particular sensor
(there will be more than one of each type of sensor later)

Key point: multiple sensors work. Tried it with temperature and light sensor
but the code for the humidity sensor is already in as well. The plan is to have
all three ready for the demo.

Current the Makefile needs more improvement: I made it a little nicer for our
needs (notice you can specify the number of each sensor on compile time with
LIGHT/DHT/TEMP_SENSOR_COUNT=val). The current issue is that my if statements
for checking if there is a zero in any of the sensor counts isn't working for
some reason, maybe I just need some sleep?

Finally, as I said before, READ is the only function that has the expected
input and output so far. This is to a degree purely syntactical so it's not a
big deal. Regardless, I plan to have the normal sensor polling functionality
syntactically correct (ie: get light sensor 1 with L1? and the radio will send
L1=val)