Added status command for actuators
status can be polled but not set.
status is automatically determined by the actions being done
to a particular actuator (whether it is being turned on/off at any time).

Status command is in format:
ATx? : get status <0/1> of actuator x

got rid of colons in commands other than I? (Pretty sure but not 100%)

increased speed of updating actuators (it seemed to be a little slow)