multi-pin fix is working, mem issues mostly done
the system is usable again but a good amount of changes have been made:

PROGMEM is used in such a way to not require it as a standard but it is HIGHLY
RECOMMENDED instead of the use of any constants

All overriden functions must print to uart, main no longer handles this as all
functions now return void

The goal anyway was to allow for multiple pins to be assigned to a module as
needed. You can have anywhere from 1 to 8 pins per module now. Considering
allowing a module not to specify any pins but there are no plans to allow more
than 8 as it wouldn't be neccessary anyway.

Need to integrate temp_sensor again now that some things have changed. While
this shouldn't take long, I have done a lot and I believe that it is time for me
to: commit and be done for the day!

This branch will die soon. It shall be rebased onto master