Started adding necessary commands for FONA, uart
Uart needs some additional functionality which is detailed with TODOs

FONA module is a good start. Have a decent idea of what commands needed
for TCP server and transparent modes. Time to see what general command are
needed as well
Will need to remake the way uart is initialized because a baudrate should be
picked, not hard-coded. Then again, not sure what to use for the FONA in that
regard.